requirements.txt
fastapi
uvicorn
python-multipart
cloudinary
pillow
facenet-pytorch
torch
opencv-python
numpy
